cleanup unused quic code
