+++ quic-forward-proxy/src/lib.rs
@@ -4,12 +4,12 @@ mod quic_util;
 pub mod tls_config_provider;
 pub mod proxy;
 pub mod validator_identity;
-mod proxy_request_format;
+pub mod proxy_request_format;
 mod cli;
 mod test_client;
 mod util;
 mod tx_store;
-mod quic_connection_utils;
+mod tpu_quic_connection_utils;
 mod quinn_auto_reconnect;
 mod outbound;
 mod inbound;
